Squashed commit of the following:
commit fa8fd46d3e4d943dce2db0f61247486c91e48294
Author: Raymond Wright <rayw1710@gmail.com>
Date:   Fri Oct 8 10:43:52 2021 +0100

    Update Qt location for Mac

commit 7d3dd8915888ca0d0bb0ba68cc6ef1fe72e6d1ad
Author: Raymond Wright <rayw1710@gmail.com>
Date:   Fri Oct 8 10:26:35 2021 +0100

    Disable cognitive complexity check for tests

    The measure digs into the internals of Catch's macros, making it
    worthless.

commit bb799f89a68713b7aa0de34f2ca1a7cd98178ebc
Author: Raymond Wright <rayw1710@gmail.com>
Date:   Fri Oct 8 10:12:12 2021 +0100

    Ignore macros for cognitive complexity

    Otherwise, the score for the tests balloons out of control thanks to
    Catch's use of macros.

commit 07629fe45719cf25e3bd0447a56504c4aa697c13
Author: Raymond Wright <rayw1710@gmail.com>
Date:   Fri Oct 8 09:47:50 2021 +0100

    Address clang-tidy complaints with song.cpp

commit cc8dd15c194b894b715e5001e6c26c547bff981b
Author: Raymond Wright <rayw1710@gmail.com>
Date:   Fri Oct 8 08:56:12 2021 +0100

    Break off difficulty/instrument conversions

commit d6dfa64e719850f9f60087c98093318797bb1e8a
Author: Raymond Wright <rayw1710@gmail.com>
Date:   Fri Oct 8 08:37:06 2021 +0100

    Fix bug introduced in refactoring

commit 9e733868994baa036e2665684d04309363de7a98
Author: Raymond Wright <rayw1710@gmail.com>
Date:   Fri Oct 8 08:16:55 2021 +0100

    Fix declaration that broke compilation

commit 743cee4e66c799cce879d062cfb1d6811ab67fef
Author: Raymond Wright <rayw1710@gmail.com>
Date:   Fri Oct 8 08:12:06 2021 +0100

    Fix clang-tidy complaint with processed.cpp

commit 041abd7a404380903e56ef9eeb6caa47983787a7
Author: Raymond Wright <rayw1710@gmail.com>
Date:   Fri Oct 8 07:51:04 2021 +0100

    Add missing include

commit 5b9d88fb4cdfacad1825a0e0243cb2aa2332e6a5
Author: Raymond Wright <rayw1710@gmail.com>
Date:   Fri Oct 8 07:39:02 2021 +0100

    Fix some clang-tidy complaints

commit bcf38cbcbb5fe81e58ffd15c3cf2734f3a8e4727
Author: Raymond Wright <rayw1710@gmail.com>
Date:   Fri Oct 8 06:13:33 2021 +0100

    Update Qt versions used on AppVeyor

+++ src/points.cpp
@@ -183,13 +183,13 @@ points_from_track(const NoteTrack<T>& track, const TimeConverter& converter,
                          return x.position.beat < y.position.beat;
                      });
 
-    auto combo = 0U;
+    auto combo = 0;
     for (auto& point : points) {
         if (!point.is_hold_point) {
             ++combo;
         }
-        const auto multiplier = std::min(
-            combo / 10 + 1, static_cast<unsigned int>(engine.max_multiplier()));
+        const auto multiplier
+            = std::min(combo / 10 + 1, engine.max_multiplier());
         point.value *= multiplier;
     }
